+++ examples/storybook/stories/Components/ExampleApplication.tsx
@@ -9,6 +9,7 @@ import editorSourceCode from '../Components/EditorWrapper?raw';
 import viewerSourceCode from '../Components/ViewerWrapper?raw';
 import styles from './styles.scss';
 import type { DraftContent, RicosTheme } from 'ricos-editor';
+import type { AvailableExperiments } from 'ricos-types';
 
 const ExampleApplication: FunctionComponent<{
   initialState?: DraftContent;
@@ -16,7 +17,8 @@ const ExampleApplication: FunctionComponent<{
   display?: 'Editor' | 'Viewer' | 'Both';
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   editorProps?: Record<string, any>;
-}> = ({ initialState, theme, display = 'Both', editorProps = {} }) => {
+  experiments?: AvailableExperiments;
+}> = ({ initialState, theme, display = 'Both', editorProps = {}, experiments = {} }) => {
   const [content, setContent] = useState(initialState);
   const showEditor = useMemo(() => display === 'Both' || display === 'Editor', [display]);
   const showViewer = useMemo(() => display === 'Both' || display === 'Viewer', [display]);
@@ -29,6 +31,7 @@ const ExampleApplication: FunctionComponent<{
             content={content}
             theme={{ ...theme, parentClass: styles['rce-wrapper'] }}
             onChange={setContent}
+            experiments={experiments}
             {...editorProps}
           />
         </RichContentEditorBox>