+++ examples/storybook/stories/Theme/ThemeStory.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Layout, Palette, ToggleSwitch } from 'wix-style-react';
 import { Page, Section, ContentState } from '../Components/StoryParts';
-import exapmleState from '../../../../e2e/tests/fixtures/storybook-example-app.json';
+import exampleState from '../../../../e2e/tests/fixtures/storybook-example-app-without-vertical-embed.json';
 import { wixPalettes, ricosPalettes } from '../../src/shared/resources/palettesExample';
 import { FONTS } from '../../src/shared/resources/fontsExample';
 import ExampleApplication from '../Components/ExampleApplication';
@@ -9,6 +9,7 @@ import { SelectorCell } from './SelectorCell';
 import type { RicosTheme } from 'ricos-common';
 import { withWixStyle } from './wixStyle';
 import { withOneAppStyle } from './oneAppStyle';
+import { ToggleTiptapButton } from '../Components/ToggleTiptapButton';
 
 const palettes = Object.keys(wixPalettes);
 
@@ -18,6 +19,7 @@ const ThemeSelector = () => {
   const [isFallback, setFallback] = useState(false);
   const [isFloatingBM, setFloatingBM] = useState(false);
   const [isOneApp, setOneApp] = useState(false);
+  const [isTiptap, setIsTiptap] = useState(false);
   const fallbackColor = isFallback ? '#FF0000' : undefined;
   const palette = ricosPalettes[palettePage];
   const values = Object.values(palette);
@@ -79,27 +81,29 @@ const ThemeSelector = () => {
 
       <div>
         <ToggleSwitch checked={isFallback} onChange={({ target }) => setFallback(target.checked)} />
-        <span>Use RED fallback color</span>
+        <span> Use RED fallback color</span>
       </div>
       <div>
         <ToggleSwitch
           checked={isFloatingBM}
           onChange={({ target }) => setFloatingBM(target.checked)}
         />
-        <span>Use BM Blue floating action color</span>
+        <span> Use BM Blue floating action color</span>
       </div>
       <div>
         <ToggleSwitch checked={isOneApp} onChange={({ target }) => setOneApp(target.checked)} />
-        <span>Use OneApp styles</span>
+        <span> Use OneApp styles</span>
       </div>
+      <ToggleTiptapButton isTiptap={isTiptap} setIsTiptap={setIsTiptap} />
       <div style={{ backgroundColor: palette.bgColor, padding: 4 }}>
         <ExampleApplication
           key={palettePage}
-          initialState={exapmleState}
+          initialState={exampleState}
           theme={createTheme({
             palette: { ...palette, fallbackColor },
             customStyles: FONTS[fontPage],
           })}
+          experiments={{ tiptapEditor: { enabled: isTiptap } }}
         />
       </div>
     </>
@@ -126,7 +130,7 @@ export default () => {
         </a>
         <ThemeSelector />
         <Section title="Content State">
-          <ContentState json={exapmleState} />
+          <ContentState json={exampleState} />
         </Section>
       </Page>
     </>